+++ requirements.txt
@@ -14,12 +14,14 @@ cryptography==37.0.2
 decorator==5.1.1
 defusedxml==0.7.1
 Deprecated==1.2.13
+distlib==0.3.5
 Django==4.0.5
 django-allauth==0.51.0
 django-cors-headers==3.12.0
 django-debug-toolbar==3.4.0
 django-extensions==3.1.5
 django-filter==21.1
+django-redis==5.2.0
 django-redis-cache==3.0.1
 django-rest-auth==0.9.5
 django-rest-authtoken==2.1.4
@@ -27,6 +29,7 @@ django-rest-framework==0.1.0
 djangorestframework==3.13.1
 djangorestframework-simplejwt==5.2.0
 executing==0.8.3
+filelock==3.7.1
 gunicorn==20.1.0
 idna==3.3
 install==1.3.5
@@ -41,10 +44,12 @@ parso==0.8.3
 pexpect==4.8.0
 pickleshare==0.7.5
 Pillow==9.1.1
+platformdirs==2.5.2
 prompt-toolkit==3.0.29
 ptyprocess==0.7.0
 pure-eval==0.2.2
 pycparser==2.21
+pycryptodome==3.15.0
 Pygments==2.12.0
 PyJWT==1.7.1
 PyMySQL==1.0.2
@@ -62,5 +67,6 @@ sqlparse==0.4.2
 stack-data==0.2.0
 traitlets==5.2.2.post1
 urllib3==1.26.9
+virtualenv==20.16.2
 wcwidth==0.2.5
 wrapt==1.14.1
